+++ spec/unit/resource/api800/c7000/firmware_driver_spec.rb
@@ -7,14 +7,12 @@
 # Unless required by applicable law or agreed to in writing, software distributed
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations under the License.
+# specific language governing permissions and limitations under the License.require 'spec_helper'
 
-require 'spec_helper'
-
-RSpec.describe OneviewSDK::API600::Synergy::FirmwareDriver do
+RSpec.describe OneviewSDK::API800::C7000::FirmwareDriver do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API300::Synergy::FirmwareDriver' do
-    expect(described_class).to be < OneviewSDK::API300::Synergy::FirmwareDriver
+  it 'inherits from OneviewSDK::API600::C7000::FirmwareDriver' do
+    expect(described_class).to be < OneviewSDK::API600::C7000::FirmwareDriver
   end
 end