+++ spec/unit/resource/api800/synergy/firmware_driver_spec.rb
@@ -11,10 +11,10 @@
 
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API600::Synergy::FirmwareDriver do
+RSpec.describe OneviewSDK::API800::Synergy::FirmwareDriver do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API300::Synergy::FirmwareDriver' do
-    expect(described_class).to be < OneviewSDK::API300::Synergy::FirmwareDriver
+  it 'inherits from OneviewSDK::API600::Synergy::FirmwareDriver' do
+    expect(described_class).to be < OneviewSDK::API600::Synergy::FirmwareDriver
   end
 end